Rename everything to be singular, e.g. generators --> generator
