+++ src/consumer.rs
