+++ src/control/mod.rs
@@ -1,3 +1,7 @@
+//! Control the audio out via inputs, branching, and many other means.
+// TODO: better docstring
+
 pub mod pot;
 pub mod mux;
 pub mod flow;
+pub mod key;
